requirements.txt
anthropic>=0.39.0
structlog>=24.1.0
pyyaml>=6.0.1
python-dotenv>=1.0.1

# Dev dependencies
pytest>=8.0.0
pytest-cov>=5.0.0
ruff>=0.5.0
mypy>=1.10.0
